Fixed #76: phet tasks for worker
Using similar structure as for MWOffliner:
- A task launching a phet container then uploading files
- An operation to pull and start the phet container

Adapted validators as there flags is not tied to mwoffliner anymore.
Added some tests.

Worth nothing that while flags are now validated based on the requested task
all other fields are flexible (image, for instance).
We shall discuss how much we want to enforce rules on config.