restore selfish mode (#407)
selfish mode used to work by filtering-out tasks on the worker itself.
the backend was not aware of the selfishness of a worker.

when switching to advanced scheduling, we started returning a single, pre-selected
task for each poll. this rendered selfish useless.

this restores selfish mode by informing the API of the selfishness on check-in.
we shall now be able to distinguish selfish workers in the UI (not impl here).
scheduler now assigns tasks according to selfish mode.