Issues #356 #411: more retries on docker API
increased number of retries on the docker API which also applies to docker hub.
retries are incrementaly longer (10s * attempt) covering a larger range of scenarios
With this, a docker API and or docker Hub issue would be attempted up to 9mn.
That should get us through all hickups and fail only due to more serious problems