consider ZIMFARM_DISK as reserved disk space (#428)
`ZIMFARM_DISK` config variable was used to set a maximum amount of disk space that
the worker could use: ie. scheduler would use this value to select doable task for it.

Because non-dedicated workers might consume disk space outside the zimfarm, we are
now considering the ZIMFARM_DISK variable to represent a dedicated amount of disk
storage for the zimfarm.

Because it's now supposed to be dedicated, should the available disk space not match it
the worker would exit, providing the reason (and the sizes concerned).

Config file comment and README have been updated accordingly.