Issue #404: added support for UI-hidden flags
Offliner flag with an extra `secret` property set to `true` would be considered
a text string which value should not be displayed on Recipe config.

In both the config table and the sample docker command, the value will be replaced
with ****.

Note that the API still returns actual value, as well as executed command from the
Debug tab and potentially logs.