Fixed task-worker loop on files-upload
main task-worker loop (`run()`) is basically doing nothing while the scraper is running.
It must still wait for the scraper to exit and upload ZIM files as those are created.
The ZIM creation detection (`.upload_files()`) is called every minute to do so.

As reported by Kevin McMurtrie in #zimfarm, the `last_check` variable used to calculate
than a minute has passed was actually never updated.
This resulted in a no-delay loop to `upload_files()`.

Fixing this might improve performances a bit as `upload_files()` does two things all the time:

1. list files on the target folder
2. query the status of the scraper container