Use x_UPLOAD_URI to specify upload path
`ZIM_UPLOAD_URI` and `LOGS_UPLOAD_URI` were specified as scheme://netloc and the worker
used to hard-code a *folder* (`zim/` or `logs/`) at upload time to place file into.
This was inherited by previous warehouse-based upload to ensure we're uploading
only to safe places.
This is not flexible enough as it forces path on server for stuff that shouldn't require it.
Full path is now specified on the ENV so we moved from.