Receive scraper-log uploaded event
We used to send the scraper log-file name on container-started as that's when we knew
its name on the file system (as we were uploading the raw docker JSON-log file).

Now that we don't want to directly upload this file, it makes no sense to use this
uncorrelated filename.
Also, now that we've moved to upload the scraper log after the scraper is completed,
uploading its name was misleading and would lead users to attempt to download
a non-uploaded file.

Now introducing a new `update` event whose purpose is just to update task data
without recording an event.