Fixed #611: better token invalidity management
When the backend (dispatcher) is restarted, all issued tokens become invalid
because we salt out tokens using a random string generated on startup.

In the UI, since we keep track of token expiration date and permission inside our cookie
we have no idea our token became invalid until we attempt to use it and get rejected.
This lead to a 401 Forbidden error message being displayed on restricted actions.

This doesn't change the backend behavior but improves the user experience in the UI:
Most API requests' errors goes through a common function that displays the error msg
If the API response was a 401 status: attempt to renew token and inform user.

New bahavior now would be to be redirected to home page with a message saying that
the token has been refreshed. User is still logged-in and can perform restricted
actions.

Further step could be to transparently retry the failing query and not be affected
at all by invalidated tokens but that would require additional work to prevent request
loops in case we're in a different scenario.