Added multi-API support to manager
A Zimfarm worker can now be attached to multiple Zimfarm backend APIs to requested to
and execute tasks for.

This new capability is transparent to the task-worker which continues to work as before.
Its only *change* is that we now always start it with its --webapi-url while we used
to rely on a default value (option was already present).

The new logic resides in the manager which now looks for its endpoints on environ
`WEB_API_URIS` (notice the **S**), a `,` separated list.

All endpoints are tested and must authenticate properly on manager startup.

- All endpoints are queried one after the other.
- The first proposed (matching) task is accepted and started.
- Polling is interrupted for 3mn upon each task start, in order to let the task
resources be assigned before polling again (previous behavior)
- Polling then resumes to next endpoint.
- Polling is still only trigger every {interval} (3mn by default) when there is no task

zmq-based event system has been removed from manager as it has never been used due to
an infrastructure issue.