Fixed #680: include worker name in running_tasks
With introduction of per-platform, cross-workers limit, a new list of all_running_tasks
is being used, from which is derived a list of the worker's tasks.

Because that large list comes from a function that used to be called for a specific worker
it did not include the worker's name in the returned tasks.
This made the filter-using per-worker list empty all the time, letting the scheduler
think it is currently not running other task when comparing platforms