Fixed #755: retrieve image tags from ghcr.io
In an ideal world, we'd dynamically query the tags list from the hosting service.
As we are leaving docker.io and only use ghcr.io for now, this only changes the
hard-coded code to query ghcr.io

Retrieving the list of tags for an image is not part of the registry API.
On Docker hub, it's part of the HTTP API that GHCR doesn't implement.
There's an equivalent endpoint though but it requires (trivial) authentication
preventing non-CORS usage so it can't run in the frontend.